requirements.txt
torch
transformers
peft
sentence-transformers
faiss-cpu
numpy
